requirements.txt
django==5.2
djangorestframework==3.16.0
google-genai
pillow
python-pptx